requirements.txt
# Core dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.0
pydantic-settings>=2.0.3
python-multipart>=0.0.6
requests>=2.31.0
numpy>=1.24.4
scipy>=1.11.2
matplotlib>=3.7.3
pandas>=2.1.0
pillow>=10.0.1

# Machine learning and deep learning
torch>=2.0.1
transformers>=4.33.2
scikit-learn>=1.3.0
umap-learn>=0.5.3

# Protein structure tools
biopython>=1.81
alphafold-prep>=0.0.12  # Utilities for preparing AlphaFold inputs
pymol-open-source>=2.5.0  # Optional, for structure visualization

# Data processing
gdown>=4.7.1  # For downloading model files from Google Drive
tqdm>=4.66.1
joblib>=1.3.2

# Monitoring and performance
prometheus-client>=0.17.1
python-json-logger>=2.0.7

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0
httpx>=0.25.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.2